Resize operation was not preserving all relevant landscape settings
Now uses CopySharedProperties to copy the bulk of the settings over.
Advantage is that this should automatically pick up any new settings properties we add in the future.

I added additional copies to handle the remaining settings.  In the future we might want to consider converting these to be shared properties as well.  Or if not shared, at least flag them as settings properties, so we can include them in the automatic copy here.

#rb jonathan.bard, luc.eygasier
#jira UE-173099

[CL 26200003 by chris tchou in 5.3 branch]