Chaos Cloth: allow authoring bending constraints with flat/arbitrary rest shape.
Updated TriangleMesh to produce a consistent ordering of indices in GetUniqueAdjacentElements so that the resulting angle calculation signs are consistent. Convention is such that the angles calculated by the bending element constraints are 0 when flat, positive when folding away from the normals direction, negative when folding toward the normal direction.

Added 3 options:
* Default is use the 3D mesh, which preserved existing behavior.
* Use the "FlatnessRatio" parameter which is 0-1, 0 = use 3D mesh angle, 1 = use flat (0) rest angle. Lerp between those values
* Use the "RestAngle" parameter which just sets the rest angle directly. As suggested by Kriss, using the absolute min of the two vertex values, which allows us to create sharp creases.

Added debug visualization of the rest angles. Mostly was helpful to verify signs were correct when reordering GetUniqueAdjacentElements.

#rb kriss.gossart

[CL 26200008 by Alex McAdams in 5.3 branch]