ChaosCloth: Replace "DeleteRenderMesh" with "DeleteElement" dataflow node. All types of elements can now be deleted, as well as the entire sim mesh or render mesh. Currently can delete individual elements by manually inputting the indices in an array. Selection support will be added in the future.
The cloth facade methods for deleting elements was added in a previous CL. This just calls those methods. Note that elements that reference the deleted element will still exist (e.g., if you delete a 2D vertex, the associated 3D vertex will still exist, but point to INDEX_NONE in lookups). This is all cleaned up in the terminal node when constructing the final asset.

#rb kriss.gossart

[CL 26220772 by Alex McAdams in 5.3 branch]