Skeletal Mesh Reduction.  Improve attribute interpolation during simplification.
Previously,  when an edge is collasped and two triangles are replaced by a single vertex, the quality of the attributes on that new vertex (e.g. UVs ) was limited for the special case when one of the triangles is on a plane that intersects the origin.  This change fixes that special case.

#jira UE-191560
#rb alexis.matte

[CL 26719158 by David Hill in 5.3 branch]