USD: Fix root layer payloads being broken when exporting USD Stage Actors with root layers relative to the project folder.
#jira UE-191121

Hopefully self-explanatory. This is why it wasn't working for Bill: He had his USD stage inside the project folder, which I never ever do.

Do you think this is good for 5.3? It's not really a "crash" or anything like that and it's workaroundable (just put your layer outside the project folder), but then again the fix is so simple... It would be a shame to ship a feature that silently breaks for users that happen to put their layers inside their project folders.

[CL 26963712 by daniel coelho in 5.3 branch]